Move event-related code to event.go
